Moved animations out of values and on to root in package
This is done to separate the API that has cross platform requirements with implementations that uses these APIs. The animation module uses the Values api but does not itself contain any specific platform code.
