+++ package/src/animation/decay/decay.ts
