+++ package/src/animation/decay/index.ts
