+++ package/src/animation/functions/index.ts
