+++ package/src/animation/functions/interpolate.ts
