+++ package/src/animation/index.ts
