+++ package/src/animation/spring/Spring.ts
