+++ package/src/animation/spring/functions/index.ts
