+++ package/src/animation/spring/functions/spring.ts
