+++ package/src/animation/spring/runSpring.ts
@@ -1,12 +1,14 @@
-import type { SkiaValue, SkiaAnimation } from "../../types";
+import type { SkiaValue, SkiaAnimation } from "../../values/types";
 import type {
   AnimationParams,
-  TimingConfig,
+  SpringConfig,
   AnimationCallback,
 } from "../types";
+import { runTiming } from "../timing/runTiming";
 
-import { getResolvedParams } from "./functions";
-import { createTiming } from "./createTiming";
+import { Spring } from "./Spring";
+import { createSpringEasing } from "./functions/spring";
+
 /**
  * Creates a new animation on an existing value that will be driven by
  * an animation value. The value will be run from / to the value in
@@ -16,18 +18,20 @@ import { createTiming } from "./createTiming";
  *
  * @param value The value to animate
  * @param toOrParams To value or Animation parameters
- * @param config Spring or timing configuration
+ * @param config Spring configuration
  * @returns an animation value that can be used to start/stop
  * the animation.
  */
-export const runTiming = (
+export const runSpring = (
   value: SkiaValue<number>,
   toOrParams: number | AnimationParams,
-  config?: TimingConfig,
+  config?: SpringConfig,
   callback?: AnimationCallback
 ): SkiaAnimation => {
-  const resolvedParameters = getResolvedParams(toOrParams, config);
-  const animation = createTiming(resolvedParameters, value, callback);
-  value.animation = animation;
-  return animation;
+  return runTiming(
+    value,
+    toOrParams,
+    createSpringEasing(config ?? Spring.Config.Default),
+    callback
+  );
 };
