+++ package/src/animation/timing/Easing.ts
