+++ package/src/animation/timing/functions/__tests__/timing.spec.ts
