+++ package/src/animation/timing/functions/bezier.ts
