+++ package/src/animation/timing/functions/getResolvedParams.ts
