+++ package/src/animation/timing/functions/index.ts
