+++ package/src/animation/timing/functions/types.ts
