+++ package/src/animation/timing/index.ts
