+++ package/src/animation/types.ts
