+++ package/src/values/index.ts
@@ -1,4 +1,3 @@
 export * from "./types";
 export * from "./hooks";
-export * from "./animation";
 export * from "./api";
